구조변경 : 액티비티와 프래그먼트 분리
splash에서 아이디 최초 실행 여부 판단함
