+++ app/src/main/res/navigation/nav_first_graph.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_first_graph"
+    app:startDestination="@id/splashFragment">
+
+    <fragment
+        android:id="@+id/splashFragment"
+        android:name="com.nbcam_final_account_book.persentation.firstpage.splash.SplashFragment"
+        android:label="first_splash_fragment"
+        tools:layout="@layout/first_splash_fragment" >
+        <action
+            android:id="@+id/action_splashFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.nbcam_final_account_book.persentation.firstpage.login.LoginFragment"
+        android:label="first_login_fragment"
+        tools:layout="@layout/first_login_fragment" />
+</navigation>
